Separate Android tests from secondary integration tests
Splits Android integration tests into their own package since the secondary
suite can take up to ~1h in CI. This should reduce max time spent in CI.
